# docs/swagger.yaml
basePath: /
definitions:
  models.NlpReplyModel:
    properties:
      distance:
        type: number
      intent:
        type: string
      keyword:
        type: string
    type: object
host: localhost:9000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /v1/nlp/record:
    delete:
      consumes:
      - application/json
      parameters:
      - description: shopID
        in: query
        name: shop_id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Drop nlp record by shop id
      tags:
      - customer
    get:
      consumes:
      - application/json
      parameters:
      - description: shopID
        in: query
        name: shop_id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Read nlp record by shop id
      tags:
      - customer
  /v1/nlp/record/reply:
    get:
      consumes:
      - application/json
      parameters:
      - description: shopID
        in: query
        name: shop_id
        required: true
        type: string
      - description: Keyword
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NlpReplyModel'
      summary: Read nlp model by shop id
      tags:
      - customer
  /v1/nlp/record/upload.xlsx:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: this is a test file
        in: formData
        name: xlsx
        required: true
        type: file
      - description: shopID
        in: query
        name: shop_id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Upload nlp record with xlsx
      tags:
      - customer
swagger: "2.0"